move external view and url to admin
