+++ src/openzaak/config/templates/admin/config_internal.html
@@ -16,7 +16,7 @@
 
 {% block content %}
 <h1>{% trans 'Service configuratie: Open Zaak' %}</h1>
-{% if request.GET.wizard == 'true' %}<h3>Step 2 of 2</h3>{% endif %}
+{% if request.GET.wizard == 'true' %}<h3>Step 2 of 3</h3>{% endif %}
 
 <div id="content-main">
     <form method="post" enctype="multipart/form-data">
