+++ src/openzaak/config/templates/admin/config_nlx.html
@@ -15,7 +15,7 @@
 
 {% block content %}
 <h1>{% trans 'Service configuratie: NLX' %}</h1>
-{% if request.GET.wizard == 'true' %}<h3>Step 1 of 2</h3>{% endif %}
+{% if request.GET.wizard == 'true' %}<h3>Step 1 of 3</h3>{% endif %}
 
 <div id="content-main">
     <form method="post" enctype="multipart/form-data">
