--- src/openzaak/config/templates/config/config_external.html
@@ -1,47 +1,0 @@
-{% extends "master.html" %}
-{% load i18n sniplates static %}
-
-{% block content %}
-
-<div class="px-3 py-3 pt-md-5 pb-md-4 mx-auto text-center">
-    <h1 class="display-4">Service configuratie: External Services</h1>
-    <p class="lead">Step 2 of 2</p>
-</div>
-
-<div class="container">
-    {{ auth_types|json_script:"auth-type-choices" }}
-    {{ nlx_choices|json_script:"nlx-choices" }}
-    {{ nlx_outway|json_script:"nlx-outway" }}
-    {{ formset_config|json_script:"formset-config" }}
-    {{ formdata|json_script:"formdata" }}
-
-    <form method="post" enctype="multipart/form-data">
-        {% csrf_token %}
-
-        {% if formset.errors %}
-            <p>{% trans "Please correct the error(s) below." %}</p>
-            {% if formset.non_form_errors %}
-            <ul>
-                {% for error in formset.non_form_errors %}<li>{{ error }}</li>{% endfor %}
-            </ul>
-            {% endif %}
-        {% endif %}
-
-        <div id="react-external-services">
-{#             empty on purpose - react managed #}
-        </div>
-
-
-        <div class="btn-toolbar justify-content-between py-sm-3">
-            <a class="btn btn-danger" href="{% url 'config-detail' %}">{% trans "cancel" %}</a>
-            <input class="btn btn-success" type="submit" name="submit" value="{% trans "submit" %}" />
-        </div>
-    </form>
-
-</div>
-{% endblock %}
-
-
-{% block extra_scripts %}
-    <script src="{% static 'js/main.js' %}"></script>
-{% endblock %}
